scraper/requirements.txt
Flask==1.0.2
newspaper3k==0.2.8